+++ CHANGELOG.md
@@ -69,6 +69,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - Add assemblies to AutoMapper service injection
 - Household import to use the job system
 - Reporting SQL IN clause error
+- Badge image dimension restrictions
 
 ### Fixed
 - Instance name enrichment for logging